requirements.txt
python-telegram-bot>=20.0
gpxpy>=1.5.0
requests>=2.28.0
